+++ test/TestVariable.cpp
@@ -14,368 +14,8 @@ fuzzylite (R), a fuzzy logic control library in C++.
  fuzzylite is a registered trademark of FuzzyLite Limited.
  */
 
-#include "Headers.h"
-
-namespace fuzzylite { namespace test {
-    // curiously recurring template pattern (CRTP): https://en.wikipedia.org/wiki/Curiously_recurring_template_pattern
-    template <typename DerivedAssert, typename T>
-    struct Assert {
-        std::unique_ptr<T> variable;
-        Assert() = default;
-
-        explicit Assert(std::unique_ptr<T> actual) : variable(std::move(actual)) {}
-
-        explicit Assert(T* actual) : variable(actual) {}
-
-        DerivedAssert& self() {
-            return static_cast<DerivedAssert&>(*this);
-        }
-
-        auto& has_name(const std::string& name) {
-            CHECK(variable->getName() == name);
-            return self();
-        }
-
-        auto& has_description(const std::string& description) {
-            CHECK(variable->getDescription() == description);
-            return self();
-        }
-
-        auto& is_enabled(bool enabled = true) {
-            CHECK(variable->isEnabled() == enabled);
-            return self();
-        }
-
-        auto& has_type(Variable::Type type) {
-            CHECK(variable->type() == type);
-            return self();
-        }
-
-        auto& has_range(scalar minimum, scalar maximum) {
-            CHECK_THAT(variable->getMinimum(), Approximates(minimum));
-            CHECK_THAT(variable->getMaximum(), Approximates(maximum));
-            return self();
-        }
-
-        auto& has_value(scalar value) {
-            CHECK_THAT(variable->getValue(), Approximates(value));
-            return self();
-        }
-
-        auto& locks_value_in_range(bool locks = true) {
-            CHECK(variable->isLockValueInRange() == locks);
-            return self();
-        }
-
-        auto& exports_fll(const std::string& fll) {
-            CHECK(variable->toString() == fll);
-            return self();
-        }
-
-        auto& exports_fll(const std::vector<std::string>& expected) {
-            auto obtained = Op::split(variable->toString(), "\n");
-            for (std::size_t i = 0; i < obtained.size(); ++i)
-                obtained[i] = Op::trim(obtained[i]);
-            CHECK(obtained == expected);
-            return self();
-        }
-
-        auto& fuzzify(scalar value, const std::string& expected) {
-            auto obtained = variable->fuzzify(value);
-            CHECK(obtained == expected);
-            return self();
-        }
-
-        auto& fuzzy_values(std::vector<std::pair<scalar, std::string>> values_expected) {
-            for (const auto& pair : values_expected) {
-                const std::string& expected = pair.second;
-                auto obtained = variable->fuzzify(pair.first);
-                CHECK(obtained == expected);
-            }
-            return self();
-        }
-
-        auto& highest_membership(scalar x, const Activated& expected) {
-            scalar degree;
-            Term* highest = variable->highestMembership(x, &degree);
-            CHECK(highest == expected.getTerm());
-            CHECK_THAT(degree, Approximates(expected.getDegree()));
-            return self();
-        }
-
-        auto& highest_activation(scalar x, std::vector<Activated> expected) {
-            CAPTURE(x);
-            auto maxActivations = variable->maxActivations(x);
-            REQUIRE(maxActivations.size() == expected.size());
-            for (std::size_t i = 0; i < expected.size(); ++i) {
-                CHECK(maxActivations.at(i).getTerm() == expected.at(i).getTerm());
-                CHECK_THAT(maxActivations.at(i).getDegree(), Approximates(expected.at(i).getDegree()));
-            }
-            if (expected.size() == 1)
-                highest_membership(x, expected.front());
-            return self();
-        }
-
-        auto& highest_activation(const std::vector<std::pair<scalar, std::vector<Activated>>>& highestActivations) {
-            for (auto& pair : highestActivations)
-                highest_activation(pair.first, pair.second);
-            return self();
-        }
-
-        virtual DerivedAssert& equals(const T& another) {
-            CHECK(variable->getName() == another.getName());
-            CHECK(variable->getDescription() == another.getDescription());
-            CHECK(variable->isEnabled() == another.isEnabled());
-            CHECK(variable->isLockValueInRange() == another.isLockValueInRange());
-
-            CHECK_THAT(variable->getMinimum(), Approximates(another.getMinimum()));
-            CHECK_THAT(variable->getMaximum(), Approximates(another.getMaximum()));
-            CHECK_THAT(variable->getValue(), Approximates(another.getValue()));
-
-            CHECK(variable->toString() == another.toString());
-            return self();
-        }
-
-        auto& can_clone() {
-            auto clone = std::unique_ptr<T>(variable->clone());
-            return equals(*clone.get());
-        }
-    };
-
-    struct VariableAssert : Assert<VariableAssert, Variable> {
-        using Assert::Assert;  // inherit constructors
-    };
-
-    struct InputVariableAssert : Assert<InputVariableAssert, InputVariable> {
-        using Assert::Assert;  // inherit constructors
-
-        auto& has_fuzzy_value(scalar input, const std::string& expected) {
-            const scalar previousInput = variable->getValue();
-            variable->setValue(input);
-            CHECK(variable->fuzzyInputValue() == expected);
-            variable->setValue(previousInput);
-            return self();
-        }
-
-        auto& has_fuzzy_values(const std::vector<std::pair<scalar, std::string>> expected) {
-            for (auto& pair : expected)
-                has_fuzzy_value(pair.first, pair.second);
-            return self();
-        }
-
-        InputVariableAssert& equals(const InputVariable& another) override {
-            return Assert::equals(another);
-        }
-    };
-
-    struct MockDefuzzifier : Centroid {
-        scalar return_value;
-
-        explicit MockDefuzzifier(scalar return_value = nan) : return_value(return_value) {}
-
-        scalar defuzzify(const Term* term, scalar minimum, scalar maximum) const override {
-            (void)term;
-            (void)minimum;
-            (void)maximum;
-            return return_value;
-        }
-    };
-
-    class State {
-        scalar _previous, _current;
-
-      public:
-        explicit State(scalar previous = nan, scalar current = nan) : _previous(previous), _current(current) {}
-
-        State& previous(scalar value) {
-            _previous = value;
-            return *this;
-        }
-
-        State& current(scalar value) {
-            _current = value;
-            return *this;
-        }
-
-        scalar previous() const {
-            return _previous;
-        }
-
-        scalar current() const {
-            return _current;
-        }
-    };
-
-    class Combination {
-        std::string _number;
-        State _before, _after;
-        scalar _defuzzify;
-
-        enum Selector { BEFORE, AFTER } _selector = BEFORE;
-
-      public:
-        explicit Combination(std::string number) :
-            _number(number),
-            _before(State()),
-            _after(State()),
-            _defuzzify(nan) {}
-
-        std::string number() const {
-            return _number;
-        }
-
-        Combination& when() {
-            _selector = BEFORE;
-            return *this;
-        }
-
-        scalar defuzzify() const {
-            return _defuzzify;
-        }
-
-        Combination& defuzzify(scalar value) {
-            _defuzzify = value;
-            _selector = AFTER;
-            return *this;
-        }
-
-        State& state() {
-            switch (_selector) {
-                case BEFORE:
-                    return _before;
-                case AFTER:
-                    return _after;
-                default:
-                    return _before;
-            }
-        }
-
-        Combination& previous(scalar value) {
-            state().previous(value);
-            return *this;
-        }
-
-        Combination& current(scalar value) {
-            state().current(value);
-            return *this;
-        }
-
-        Combination& before(State state) {
-            _before = state;
-            return *this;
-        }
-
-        Combination& after(State state) {
-            _after = state;
-            return *this;
-        }
-
-        State before() const {
-            return _before;
-        }
-
-        State after() const {
-            return _after;
-        }
-    };
-
-    struct OutputVariableAssert : Assert<OutputVariableAssert, OutputVariable> {
-        using Assert::Assert;
-
-        auto& has_previous_value(scalar value) {
-            CHECK_THAT(variable->getPreviousValue(), Approximates(value));
-            return self();
-        }
-
-        auto& has_default_value(scalar value) {
-            CHECK_THAT(variable->getDefaultValue(), Approximates(value));
-            return self();
-        }
-
-        auto& locks_previous_value(bool locks = true) {
-            CHECK(variable->isLockPreviousValue() == locks);
-            return self();
-        }
-
-        auto& has_defuzzifier(const std::string& fll) {
-            CHECK(FllExporter().toString(variable->getDefuzzifier()) == fll);
-            return self();
-        }
-
-        auto& has_aggregation(const std::string& fll) {
-            CHECK(FllExporter().toString(variable->getAggregation()) == fll);
-            return self();
-        }
-
-        auto& when_fuzzy_output_is(const std::vector<Activated>& terms) {
-            variable->fuzzyOutput()->clear();
-            variable->fuzzyOutput()->setTerms(terms);
-            return self();
-        }
-
-        auto& when_defuzzify() {
-            variable->defuzzify();
-            return self();
-        }
-
-        auto& when_mock_defuzzify(scalar return_value) {
-            variable->setDefuzzifier(new MockDefuzzifier(return_value));
-            variable->defuzzify();
-            return self();
-        }
-
-        auto& has_fuzzy_value(const std::string& value) {
-            CHECK(variable->fuzzyOutputValue() == value);
-            return self();
-        }
-
-        OutputVariableAssert& equals(const OutputVariable& another) override {
-            Assert::equals(another);
-            CHECK_THAT(variable->getDefaultValue(), Approximates(another.getDefaultValue()));
-            CHECK_THAT(variable->getPreviousValue(), Approximates(another.getPreviousValue()));
-            CHECK(
-                FllExporter().toString(variable->getDefuzzifier()) == FllExporter().toString(another.getDefuzzifier())
-            );
-            CHECK(variable->isLockPreviousValue() == another.isLockPreviousValue());
-            return self();
-        }
-
-        OutputVariableAssert& fsm(const std::vector<Combination>& combinations) {
-            for (auto& combination : combinations) {
-                SECTION("Combination " + combination.number()) {
-                    CAPTURE(combination.number());
-                    variable->setPreviousValue(combination.before().previous());
-                    variable->setValue(combination.before().current());
-                    variable->setDefuzzifier(new MockDefuzzifier(combination.defuzzify()));
-                    variable->defuzzify();
-                    CHECK_THAT(variable->getPreviousValue(), Approximates(combination.after().previous()));
-                    CHECK_THAT(variable->getValue(), Approximates(combination.after().current()));
-                }
-            }
-            return *this;
-        }
-
-        OutputVariableAssert& fsm_deprecated_v6(const std::vector<Combination>& combinations) {
-            Constant dummy;
-            for (auto& combination : combinations) {
-                variable->fuzzyOutput()->clear();
-                SECTION("Combination " + combination.number()) {
-                    CAPTURE(combination.number());
-                    variable->setPreviousValue(combination.before().previous());
-                    variable->setValue(combination.before().current());
-                    variable->setDefuzzifier(new MockDefuzzifier(combination.defuzzify()));
-                    if (not Op::isNaN(combination.defuzzify()))
-                        variable->fuzzyOutput()->setTerms({Activated(&dummy)});
-                    variable->defuzzify_v6();
-                    CHECK_THAT(variable->getValue(), Approximates(combination.after().current()));
-                    CHECK_THAT(variable->getPreviousValue(), Approximates(combination.after().previous()));
-                }
-            }
-            return *this;
-        }
-    };
-
-}}
+#include "test/AssertVariable.h"
+#include "test/Headers.h"
 
 namespace fuzzylite { namespace test { namespace variable {
 
@@ -414,14 +54,16 @@ namespace fuzzylite { namespace test { namespace variable {
                 .has_value(0.0)
                 .has_range(-1, 1)
                 .locks_value_in_range(true)
-                .exports_fll(std::vector<std::string>{
-                    "Variable: Test",
-                    "description: A test variable",
-                    "enabled: false",
-                    "range: -1.000 1.000",
-                    "lock-range: true",
-                    "term: triangle Triangle 1.000 2.000 3.000 0.500"
-                })
+                .exports_fll(
+                    std::vector<std::string>{
+                        "Variable: Test",
+                        "description: A test variable",
+                        "enabled: false",
+                        "range: -1.000 1.000",
+                        "lock-range: true",
+                        "term: triangle Triangle 1.000 2.000 3.000 0.500"
+                    }
+                )
                 .equals(testVariable());
         }
 
@@ -436,14 +78,16 @@ namespace fuzzylite { namespace test { namespace variable {
                 .has_value(0.0)
                 .has_range(-1, 1)
                 .locks_value_in_range(true)
-                .exports_fll(std::vector<std::string>{
-                    "Variable: Test",
-                    "description: A test variable",
-                    "enabled: false",
-                    "range: -1.000 1.000",
-                    "lock-range: true",
-                    "term: triangle Triangle 1.000 2.000 3.000 0.500"
-                })
+                .exports_fll(
+                    std::vector<std::string>{
+                        "Variable: Test",
+                        "description: A test variable",
+                        "enabled: false",
+                        "range: -1.000 1.000",
+                        "lock-range: true",
+                        "term: triangle Triangle 1.000 2.000 3.000 0.500"
+                    }
+                )
                 .equals(testVariable());
         }
 
@@ -471,14 +115,16 @@ namespace fuzzylite { namespace test { namespace variable {
                 .has_value(0.0)
                 .has_range(-1, 1)
                 .locks_value_in_range(true)
-                .exports_fll(std::vector<std::string>{
-                    "Variable: Test",
-                    "description: A test variable",
-                    "enabled: false",
-                    "range: -1.000 1.000",
-                    "lock-range: true",
-                    "term: pi Constant 3.141"
-                })
+                .exports_fll(
+                    std::vector<std::string>{
+                        "Variable: Test",
+                        "description: A test variable",
+                        "enabled: false",
+                        "range: -1.000 1.000",
+                        "lock-range: true",
+                        "term: pi Constant 3.141"
+                    }
+                )
                 .equals(test_variable);
         }
 
@@ -603,16 +249,18 @@ namespace fuzzylite { namespace test { namespace variable {
     }
 
     TEST_CASE("Variable fuzzifies", "[variable]") {
-        VariableAssert(std::make_unique<Variable>(
-                           "Temperature",
-                           0,
-                           100,
-                           std::vector<Term*>{
-                               new Triangle("Low", -1.0, -1.0, 0.0),
-                               new Triangle("Medium", -0.5, 0.0, 0.5),
-                               new Triangle("High", 0.0, 1.0, 1.0)
-                           }
-                       ))
+        VariableAssert(
+            std::make_unique<Variable>(
+                "Temperature",
+                0,
+                100,
+                std::vector<Term*>{
+                    new Triangle("Low", -1.0, -1.0, 0.0),
+                    new Triangle("Medium", -0.5, 0.0, 0.5),
+                    new Triangle("High", 0.0, 1.0, 1.0)
+                }
+            )
+        )
             .fuzzy_values({
                 {-1.00, "1.000/Low + 0.000/Medium + 0.000/High"},
                 {-0.50, "0.500/Low + 0.000/Medium + 0.000/High"},
@@ -627,19 +275,21 @@ namespace fuzzylite { namespace test { namespace variable {
                 {inf, "0.000/Low + 0.000/Medium + 0.000/High"},
                 {-inf, "0.000/Low + 0.000/Medium + 0.000/High"},
             });
-        VariableAssert(std::make_unique<Variable>(
-                           "Temperature",
-                           -inf,
-                           inf,
-                           std::vector<Term*>{
-                               new Constant("A", -1),
-                               new Constant("B", 0),
-                               new Constant("C", 1),
-                               new Constant("D", -inf),
-                               new Constant("E", inf),
-                               new Constant("F", nan),
-                           }
-                       ))
+        VariableAssert(
+            std::make_unique<Variable>(
+                "Temperature",
+                -inf,
+                inf,
+                std::vector<Term*>{
+                    new Constant("A", -1),
+                    new Constant("B", 0),
+                    new Constant("C", 1),
+                    new Constant("D", -inf),
+                    new Constant("E", inf),
+                    new Constant("F", nan),
+                }
+            )
+        )
             //@todo: in pyfuzzylite, expected = "-1.000/A + 0.000/B + 1.000/C + 0.000/D + 1.000/E + 0.000/F"
             .fuzzify(nan, "-1.000/A + 0.000/B + 1.000/C - inf/D + inf/E + nan/F");
     }
@@ -686,42 +336,50 @@ namespace fuzzylite { namespace test { namespace input_variable {
                 .is_enabled()
                 .has_range(-inf, inf)
                 .locks_value_in_range(false)
-                .exports_fll(std::vector<std::string>{
-                    "InputVariable: _", "enabled: true", "range: -inf inf", "lock-range: false"
-                });
+                .exports_fll(
+                    std::vector<std::string>{
+                        "InputVariable: _", "enabled: true", "range: -inf inf", "lock-range: false"
+                    }
+                );
         }
         SECTION("Custom constructor") {
-            InputVariableAssert(std::make_unique<InputVariable>(
-                                    "Test",
-                                    -1.0,
-                                    1.0,  //
-                                    std::vector<Term*>{new Constant("A", 1), new Constant("B", 2)}
-                                ))
+            InputVariableAssert(
+                std::make_unique<InputVariable>(
+                    "Test",
+                    -1.0,
+                    1.0,  //
+                    std::vector<Term*>{new Constant("A", 1), new Constant("B", 2)}
+                )
+            )
                 .has_name("Test")
                 .has_description("")
                 .is_enabled()
                 .has_range(-1, 1)
                 .locks_value_in_range(false)
-                .exports_fll(std::vector<std::string>{
-                    {"InputVariable: Test",
-                     "enabled: true",
-                     "range: -1.000 1.000",
-                     "lock-range: false",
-                     "term: A Constant 1.000",
-                     "term: B Constant 2.000"}
-                });
+                .exports_fll(
+                    std::vector<std::string>{
+                        {"InputVariable: Test",
+                         "enabled: true",
+                         "range: -1.000 1.000",
+                         "lock-range: false",
+                         "term: A Constant 1.000",
+                         "term: B Constant 2.000"}
+                    }
+                );
         }
         SECTION("Copy constructor") {
             InputVariableAssert(std::make_unique<InputVariable>(testVariable()))
                 .equals(testVariable())
-                .exports_fll(std::vector<std::string>{
-                    "InputVariable: Test",
-                    "enabled: true",
-                    "range: -1.000 1.000",
-                    "lock-range: false",
-                    "term: A Constant 1.000",
-                    "term: B Constant 2.000"
-                });
+                .exports_fll(
+                    std::vector<std::string>{
+                        "InputVariable: Test",
+                        "enabled: true",
+                        "range: -1.000 1.000",
+                        "lock-range: false",
+                        "term: A Constant 1.000",
+                        "term: B Constant 2.000"
+                    }
+                );
         }
         SECTION("Copy assign") {
             InputVariable copy;
@@ -729,14 +387,16 @@ namespace fuzzylite { namespace test { namespace input_variable {
             copy = test;
             InputVariableAssert(std::make_unique<InputVariable>(copy))
                 .equals(testVariable())
-                .exports_fll(std::vector<std::string>{
-                    "InputVariable: Test",
-                    "enabled: true",
-                    "range: -1.000 1.000",
-                    "lock-range: false",
-                    "term: A Constant 1.000",
-                    "term: B Constant 2.000"
-                });
+                .exports_fll(
+                    std::vector<std::string>{
+                        "InputVariable: Test",
+                        "enabled: true",
+                        "range: -1.000 1.000",
+                        "lock-range: false",
+                        "term: A Constant 1.000",
+                        "term: B Constant 2.000"
+                    }
+                );
         }
 
         SECTION("Clone") {
@@ -748,16 +408,18 @@ namespace fuzzylite { namespace test { namespace input_variable {
     }
 
     TEST_CASE("InputVariable: fuzzy values") {
-        InputVariableAssert(std::make_unique<InputVariable>(
-                                "name",
-                                -1.0,
-                                1.0,
-                                std::vector<Term*>{
-                                    new Triangle{"Low", -1.0, -1.0, 0.0},
-                                    new Triangle{"Medium", -0.5, 0.0, 0.5},
-                                    new Triangle{"High", 0.0, 1.0, 1.0},
-                                }
-                            ))
+        InputVariableAssert(
+            std::make_unique<InputVariable>(
+                "name",
+                -1.0,
+                1.0,
+                std::vector<Term*>{
+                    new Triangle{"Low", -1.0, -1.0, 0.0},
+                    new Triangle{"Medium", -0.5, 0.0, 0.5},
+                    new Triangle{"High", 0.0, 1.0, 1.0},
+                }
+            )
+        )
             .has_fuzzy_values({
                 {-1.00, "1.000/Low + 0.000/Medium + 0.000/High"},
                 {-0.50, "0.500/Low + 0.000/Medium + 0.000/High"},
@@ -804,24 +466,28 @@ namespace fuzzylite { namespace test { namespace output_variable {
                 .has_default_value(nan)
                 .locks_previous_value(false)
                 .has_aggregation("none")
-                .exports_fll(std::vector<std::string>{
-                    "OutputVariable: _",
-                    "enabled: true",
-                    "range: -inf inf",
-                    "lock-range: false",
-                    "aggregation: none",
-                    "defuzzifier: none",
-                    "default: nan",
-                    "lock-previous: false"
-                });
+                .exports_fll(
+                    std::vector<std::string>{
+                        "OutputVariable: _",
+                        "enabled: true",
+                        "range: -inf inf",
+                        "lock-range: false",
+                        "aggregation: none",
+                        "defuzzifier: none",
+                        "default: nan",
+                        "lock-previous: false"
+                    }
+                );
         }
         SECTION("Custom constructor") {
-            OutputVariableAssert(std::make_unique<OutputVariable>(
-                                     "Test",
-                                     -1.0,
-                                     1.0,  //
-                                     std::vector<Term*>{new Constant("A", 1), new Constant("B", 2)}
-                                 ))
+            OutputVariableAssert(
+                std::make_unique<OutputVariable>(
+                    "Test",
+                    -1.0,
+                    1.0,  //
+                    std::vector<Term*>{new Constant("A", 1), new Constant("B", 2)}
+                )
+            )
                 .has_name("Test")
                 .has_description("")
                 .is_enabled()
@@ -831,36 +497,40 @@ namespace fuzzylite { namespace test { namespace output_variable {
                 .has_default_value(nan)
                 .locks_previous_value(false)
                 .has_aggregation("none")
-                .exports_fll(std::vector<std::string>{
-                    {"OutputVariable: Test",
-                     "enabled: true",
-                     "range: -1.000 1.000",
-                     "lock-range: false",
-                     "aggregation: none",
-                     "defuzzifier: none",
-                     "default: nan",
-                     "lock-previous: false",
-                     "term: A Constant 1.000",
-                     "term: B Constant 2.000"}
-                });
+                .exports_fll(
+                    std::vector<std::string>{
+                        {"OutputVariable: Test",
+                         "enabled: true",
+                         "range: -1.000 1.000",
+                         "lock-range: false",
+                         "aggregation: none",
+                         "defuzzifier: none",
+                         "default: nan",
+                         "lock-previous: false",
+                         "term: A Constant 1.000",
+                         "term: B Constant 2.000"}
+                    }
+                );
         }
 
         // OutputVariable
         SECTION("Copy constructor") {
             OutputVariableAssert(std::make_unique<OutputVariable>(testVariable()))
                 .equals(testVariable())
-                .exports_fll(std::vector<std::string>{
-                    "OutputVariable: Test",
-                    "enabled: true",
-                    "range: -1.000 1.000",
-                    "lock-range: false",
-                    "aggregation: Maximum",
-                    "defuzzifier: Centroid",
-                    "default: 0.000",
-                    "lock-previous: true",
-                    "term: A Constant 1.000",
-                    "term: B Constant 2.000"
-                });
+                .exports_fll(
+                    std::vector<std::string>{
+                        "OutputVariable: Test",
+                        "enabled: true",
+                        "range: -1.000 1.000",
+                        "lock-range: false",
+                        "aggregation: Maximum",
+                        "defuzzifier: Centroid",
+                        "default: 0.000",
+                        "lock-previous: true",
+                        "term: A Constant 1.000",
+                        "term: B Constant 2.000"
+                    }
+                );
         }
         SECTION("Copy assign") {
             OutputVariable copy;
@@ -868,18 +538,20 @@ namespace fuzzylite { namespace test { namespace output_variable {
             copy = test;
             OutputVariableAssert(std::make_unique<OutputVariable>(copy))
                 .equals(testVariable())
-                .exports_fll(std::vector<std::string>{
-                    "OutputVariable: Test",
-                    "enabled: true",
-                    "range: -1.000 1.000",
-                    "lock-range: false",
-                    "aggregation: Maximum",
-                    "defuzzifier: Centroid",
-                    "default: 0.000",
-                    "lock-previous: true",
-                    "term: A Constant 1.000",
-                    "term: B Constant 2.000"
-                });
+                .exports_fll(
+                    std::vector<std::string>{
+                        "OutputVariable: Test",
+                        "enabled: true",
+                        "range: -1.000 1.000",
+                        "lock-range: false",
+                        "aggregation: Maximum",
+                        "defuzzifier: Centroid",
+                        "default: 0.000",
+                        "lock-previous: true",
+                        "term: A Constant 1.000",
+                        "term: B Constant 2.000"
+                    }
+                );
         }
 
         SECTION("Clone") {
@@ -942,9 +614,11 @@ namespace fuzzylite { namespace test { namespace output_variable {
             CHECK_THROWS_MATCHES(
                 variable.defuzzify(),
                 Exception,
-                Catch::Matchers::MessageMatches(Catch::Matchers::StartsWith(
-                    "[defuzzify error] expected a defuzzifier in variable 'test', but got null"
-                ))
+                Catch::Matchers::MessageMatches(
+                    Catch::Matchers::StartsWith(
+                        "[defuzzify error] expected a defuzzifier in variable 'test', but got null"
+                    )
+                )
             );
         }
 
@@ -1027,9 +701,11 @@ namespace fuzzylite { namespace test { namespace output_variable {
                 Combination("27").when().previous(inf).current(inf).defuzzify(inf).current(inf).previous(inf),
             });
         }
-        SECTION("(b) When default is not nan and no locks, value is defuzzified, default is used, and previous "
-                "value "
-                "is updated") {
+        SECTION(
+            "(b) When default is not nan and no locks, value is defuzzified, default is used, and previous "
+            "value "
+            "is updated"
+        ) {
             const scalar def = 0.7;
             assert_that.variable->setDefaultValue(def);
             assert_that.variable->setLockValueInRange(false);
@@ -1068,9 +744,11 @@ namespace fuzzylite { namespace test { namespace output_variable {
             });
         }
 
-        SECTION("(c) When default is higher than the range, and locks previous and value, then default is "
-                "capped when used"
-                "used and stored in previous") {
+        SECTION(
+            "(c) When default is higher than the range, and locks previous and value, then default is "
+            "capped when used"
+            "used and stored in previous"
+        ) {
             const scalar def = 7.7;
             const scalar def_cap = 1.0;
             const scalar cap = 1.0;
@@ -1111,8 +789,10 @@ namespace fuzzylite { namespace test { namespace output_variable {
             });
         }
 
-        SECTION("(d) When default is higher than the range, and locks value in range, then default is capped when "
-                "used and stored in previous") {
+        SECTION(
+            "(d) When default is higher than the range, and locks value in range, then default is capped when "
+            "used and stored in previous"
+        ) {
             const scalar def = 7.7;
             const scalar def_cap = 1.0;
             const scalar cap = 1.0;
@@ -1153,9 +833,11 @@ namespace fuzzylite { namespace test { namespace output_variable {
             });
         }
 
-        SECTION("(e) When default is higher than the range, and locks previous value, "
-                "previous is used when defuzzification is nan"
-                "default is used when previous is nan") {
+        SECTION(
+            "(e) When default is higher than the range, and locks previous value, "
+            "previous is used when defuzzification is nan"
+            "default is used when previous is nan"
+        ) {
             const scalar def = 7.7;
             assert_that.variable->setDefaultValue(def);
             assert_that.variable->setLockValueInRange(false);
@@ -1208,17 +890,23 @@ namespace fuzzylite { namespace test { namespace output_variable {
                 Combination("02").when().previous(0.0).current(0.0).defuzzify(nan).current(def).previous(0.0),
                 Combination("03").when().previous(0.0).current(0.0).defuzzify(inf).current(inf).previous(0.0),
                 // previous values won't update to non-finite current values
-                Combination("04").when().previous(0.0).current(nan).defuzzify(0.0).current(0.0).previous(0.0
+                Combination("04").when().previous(0.0).current(nan).defuzzify(0.0).current(0.0).previous(
+                    0.0
                 ),  //.previous(nan),
-                Combination("05").when().previous(0.0).current(nan).defuzzify(nan).current(def).previous(0.0
+                Combination("05").when().previous(0.0).current(nan).defuzzify(nan).current(def).previous(
+                    0.0
                 ),  //.previous(nan),
-                Combination("06").when().previous(0.0).current(nan).defuzzify(inf).current(inf).previous(0.0
+                Combination("06").when().previous(0.0).current(nan).defuzzify(inf).current(inf).previous(
+                    0.0
                 ),  //.previous(nan),
-                Combination("07").when().previous(0.0).current(inf).defuzzify(0.0).current(0.0).previous(0.0
+                Combination("07").when().previous(0.0).current(inf).defuzzify(0.0).current(0.0).previous(
+                    0.0
                 ),  //.previous(inf),
-                Combination("08").when().previous(0.0).current(inf).defuzzify(nan).current(def).previous(0.0
+                Combination("08").when().previous(0.0).current(inf).defuzzify(nan).current(def).previous(
+                    0.0
                 ),  //.previous(inf),
-                Combination("09").when().previous(0.0).current(inf).defuzzify(inf).current(inf).previous(0.0
+                Combination("09").when().previous(0.0).current(inf).defuzzify(inf).current(inf).previous(
+                    0.0
                 ),  //.previous(inf),
 
                 Combination("10").when().previous(nan).current(0.0).defuzzify(0.0).current(0.0).previous(0.0),
@@ -1228,30 +916,38 @@ namespace fuzzylite { namespace test { namespace output_variable {
                 Combination("14").when().previous(nan).current(nan).defuzzify(nan).current(def).previous(nan),
                 Combination("15").when().previous(nan).current(nan).defuzzify(inf).current(inf).previous(nan),
                 // previous values won't update to non-finite current values
-                Combination("16").when().previous(nan).current(inf).defuzzify(0.0).current(0.0).previous(nan
+                Combination("16").when().previous(nan).current(inf).defuzzify(0.0).current(0.0).previous(
+                    nan
                 ),  //.previous(inf),
-                Combination("17").when().previous(nan).current(inf).defuzzify(nan).current(def).previous(nan
+                Combination("17").when().previous(nan).current(inf).defuzzify(nan).current(def).previous(
+                    nan
                 ),  //.previous(inf),
-                Combination("18").when().previous(nan).current(inf).defuzzify(inf).current(inf).previous(nan
+                Combination("18").when().previous(nan).current(inf).defuzzify(inf).current(inf).previous(
+                    nan
                 ),  //.previous(inf),
 
                 Combination("19").when().previous(inf).current(0.0).defuzzify(0.0).current(0.0).previous(0.0),
                 Combination("20").when().previous(inf).current(0.0).defuzzify(nan).current(def).previous(0.0),
                 Combination("21").when().previous(inf).current(0.0).defuzzify(inf).current(inf).previous(0.0),
                 // previous values won't update to non-finite current values
-                Combination("22").when().previous(inf).current(nan).defuzzify(0.0).current(0.0).previous(inf
+                Combination("22").when().previous(inf).current(nan).defuzzify(0.0).current(0.0).previous(
+                    inf
                 ),  //.previous(nan),
-                Combination("23").when().previous(inf).current(nan).defuzzify(nan).current(def).previous(inf
+                Combination("23").when().previous(inf).current(nan).defuzzify(nan).current(def).previous(
+                    inf
                 ),  //.previous(nan),
-                Combination("24").when().previous(inf).current(nan).defuzzify(inf).current(inf).previous(inf
+                Combination("24").when().previous(inf).current(nan).defuzzify(inf).current(inf).previous(
+                    inf
                 ),  //.previous(nan),
                 Combination("25").when().previous(inf).current(inf).defuzzify(0.0).current(0.0).previous(inf),
                 Combination("26").when().previous(inf).current(inf).defuzzify(nan).current(def).previous(inf),
                 Combination("27").when().previous(inf).current(inf).defuzzify(inf).current(inf).previous(inf),
             });
         }
-        SECTION("(b) When default is not nan and no locks, value is defuzzified, default is used, and previous "
-                "value is updated") {
+        SECTION(
+            "(b) When default is not nan and no locks, value is defuzzified, default is used, and previous "
+            "value is updated"
+        ) {
             const scalar def = 0.7;
             assert_that.variable->setDefaultValue(def);
             assert_that.variable->setLockValueInRange(false);
@@ -1262,17 +958,23 @@ namespace fuzzylite { namespace test { namespace output_variable {
                 Combination("02").when().previous(0.0).current(0.0).defuzzify(nan).current(def).previous(0.0),
                 Combination("03").when().previous(0.0).current(0.0).defuzzify(inf).current(inf).previous(0.0),
                 // previous values won't update to non-finite values
-                Combination("04").when().previous(0.0).current(nan).defuzzify(0.0).current(0.0).previous(0.0
+                Combination("04").when().previous(0.0).current(nan).defuzzify(0.0).current(0.0).previous(
+                    0.0
                 ),  //.previous(nan),
-                Combination("05").when().previous(0.0).current(nan).defuzzify(nan).current(def).previous(0.0
+                Combination("05").when().previous(0.0).current(nan).defuzzify(nan).current(def).previous(
+                    0.0
                 ),  //.previous(nan),
-                Combination("06").when().previous(0.0).current(nan).defuzzify(inf).current(inf).previous(0.0
+                Combination("06").when().previous(0.0).current(nan).defuzzify(inf).current(inf).previous(
+                    0.0
                 ),  //.previous(nan),
-                Combination("07").when().previous(0.0).current(inf).defuzzify(0.0).current(0.0).previous(0.0
+                Combination("07").when().previous(0.0).current(inf).defuzzify(0.0).current(0.0).previous(
+                    0.0
                 ),  //.previous(inf),
-                Combination("08").when().previous(0.0).current(inf).defuzzify(nan).current(def).previous(0.0
+                Combination("08").when().previous(0.0).current(inf).defuzzify(nan).current(def).previous(
+                    0.0
                 ),  //.previous(inf),
-                Combination("09").when().previous(0.0).current(inf).defuzzify(inf).current(inf).previous(0.0
+                Combination("09").when().previous(0.0).current(inf).defuzzify(inf).current(inf).previous(
+                    0.0
                 ),  //.previous(inf),
 
                 Combination("10").when().previous(nan).current(0.0).defuzzify(0.0).current(0.0).previous(0.0),
@@ -1283,22 +985,28 @@ namespace fuzzylite { namespace test { namespace output_variable {
                 Combination("15").when().previous(nan).current(nan).defuzzify(inf).current(inf).previous(nan),
 
                 // previous values won't update to non-finite current values
-                Combination("16").when().previous(nan).current(inf).defuzzify(0.0).current(0.0).previous(nan
+                Combination("16").when().previous(nan).current(inf).defuzzify(0.0).current(0.0).previous(
+                    nan
                 ),  //.previous(inf),
-                Combination("17").when().previous(nan).current(inf).defuzzify(nan).current(def).previous(nan
+                Combination("17").when().previous(nan).current(inf).defuzzify(nan).current(def).previous(
+                    nan
                 ),  //.previous(inf),
-                Combination("18").when().previous(nan).current(inf).defuzzify(inf).current(inf).previous(nan
+                Combination("18").when().previous(nan).current(inf).defuzzify(inf).current(inf).previous(
+                    nan
                 ),  //.previous(inf),
 
                 Combination("19").when().previous(inf).current(0.0).defuzzify(0.0).current(0.0).previous(0.0),
                 Combination("20").when().previous(inf).current(0.0).defuzzify(nan).current(def).previous(0.0),
                 Combination("21").when().previous(inf).current(0.0).defuzzify(inf).current(inf).previous(0.0),
                 // previous values won't update to non-finite current values
-                Combination("22").when().previous(inf).current(nan).defuzzify(0.0).current(0.0).previous(inf
+                Combination("22").when().previous(inf).current(nan).defuzzify(0.0).current(0.0).previous(
+                    inf
                 ),  //.previous(nan),
-                Combination("23").when().previous(inf).current(nan).defuzzify(nan).current(def).previous(inf
+                Combination("23").when().previous(inf).current(nan).defuzzify(nan).current(def).previous(
+                    inf
                 ),  //.previous(nan),
-                Combination("24").when().previous(inf).current(nan).defuzzify(inf).current(inf).previous(inf
+                Combination("24").when().previous(inf).current(nan).defuzzify(inf).current(inf).previous(
+                    inf
                 ),  //.previous(nan),
                 Combination("25").when().previous(inf).current(inf).defuzzify(0.0).current(0.0).previous(inf),
                 Combination("26").when().previous(inf).current(inf).defuzzify(nan).current(def).previous(inf),
@@ -1306,9 +1014,11 @@ namespace fuzzylite { namespace test { namespace output_variable {
             });
         }
 
-        SECTION("(c) When default is higher than the range, and locks previous and value, then default is "
-                "capped when used"
-                "used and stored in previous") {
+        SECTION(
+            "(c) When default is higher than the range, and locks previous and value, then default is "
+            "capped when used"
+            "used and stored in previous"
+        ) {
             const scalar def = 7.7;
             const scalar def_cap = 1.0;
             const scalar cap = 1.0;
@@ -1321,20 +1031,25 @@ namespace fuzzylite { namespace test { namespace output_variable {
                 Combination("02").when().previous(0.0).current(0.0).defuzzify(nan).current(0.0).previous(0.0),
                 Combination("03").when().previous(0.0).current(0.0).defuzzify(inf).current(cap).previous(0.0),
                 // previous values won't update to non-finite current values
-                Combination("04").when().previous(0.0).current(nan).defuzzify(0.0).current(0.0).previous(0.0
+                Combination("04").when().previous(0.0).current(nan).defuzzify(0.0).current(0.0).previous(
+                    0.0
                 ),  //.previous(nan),
-                Combination("05").when().previous(0.0).current(nan).defuzzify(nan).current(0.0).previous(0.0
+                Combination("05").when().previous(0.0).current(nan).defuzzify(nan).current(0.0).previous(
+                    0.0
                 ),  //.previous(nan),
-                Combination("06").when().previous(0.0).current(nan).defuzzify(inf).current(cap).previous(0.0
+                Combination("06").when().previous(0.0).current(nan).defuzzify(inf).current(cap).previous(
+                    0.0
                 ),  //.previous(nan),
                 Combination("07").when().previous(0.0).current(inf).defuzzify(0.0).current(0.0).previous(cap),
                 // current = cap (won't change), previous = cap because updates
-                Combination("08").when().previous(0.0).current(inf).defuzzify(nan).current(cap).previous(cap
+                Combination("08").when().previous(0.0).current(inf).defuzzify(nan).current(cap).previous(
+                    cap
                 ),  // current(0.0),
                 Combination("09").when().previous(0.0).current(inf).defuzzify(inf).current(cap).previous(cap),
 
                 Combination("10").when().previous(nan).current(0.0).defuzzify(0.0).current(0.0).previous(0.0),
-                Combination("11").when().previous(nan).current(0.0).defuzzify(nan).current(0.0).previous(0.0
+                Combination("11").when().previous(nan).current(0.0).defuzzify(nan).current(0.0).previous(
+                    0.0
                 ),  // current(def_cap).previous(0.0),
                 Combination("12").when().previous(nan).current(0.0).defuzzify(inf).current(cap).previous(0.0),
                 Combination("13").when().previous(nan).current(nan).defuzzify(0.0).current(0.0).previous(nan),
@@ -1345,14 +1060,18 @@ namespace fuzzylite { namespace test { namespace output_variable {
                 Combination("18").when().previous(nan).current(inf).defuzzify(inf).current(cap).previous(cap),
 
                 Combination("19").when().previous(inf).current(0.0).defuzzify(0.0).current(0.0).previous(0.0),
-                Combination("20").when().previous(inf).current(0.0).defuzzify(nan).current(0.0).previous(0.0
+                Combination("20").when().previous(inf).current(0.0).defuzzify(nan).current(0.0).previous(
+                    0.0
                 ),  // current(def_cap)
                 Combination("21").when().previous(inf).current(0.0).defuzzify(inf).current(cap).previous(0.0),
-                Combination("22").when().previous(inf).current(nan).defuzzify(0.0).current(0.0).previous(inf
+                Combination("22").when().previous(inf).current(nan).defuzzify(0.0).current(0.0).previous(
+                    inf
                 ),  //.previous(nan),
-                Combination("23").when().previous(inf).current(nan).defuzzify(nan).current(def_cap).previous(inf
+                Combination("23").when().previous(inf).current(nan).defuzzify(nan).current(def_cap).previous(
+                    inf
                 ),  //.previous(nan),
-                Combination("24").when().previous(inf).current(nan).defuzzify(inf).current(cap).previous(inf
+                Combination("24").when().previous(inf).current(nan).defuzzify(inf).current(cap).previous(
+                    inf
                 ),  //.previous(nan),
                 Combination("25").when().previous(inf).current(inf).defuzzify(0.0).current(0.0).previous(cap),
                 Combination("26").when().previous(inf).current(inf).defuzzify(nan).current(def_cap).previous(cap),
@@ -1360,8 +1079,10 @@ namespace fuzzylite { namespace test { namespace output_variable {
             });
         }
 
-        SECTION("(d) When default is higher than the range, and locks value in range, then default is capped when "
-                "used and stored in previous") {
+        SECTION(
+            "(d) When default is higher than the range, and locks value in range, then default is capped when "
+            "used and stored in previous"
+        ) {
             const scalar def = 7.7;
             const scalar def_cap = 1.0;
             const scalar cap = 1.0;
@@ -1373,11 +1094,14 @@ namespace fuzzylite { namespace test { namespace output_variable {
                 Combination("01").when().previous(0.0).current(0.0).defuzzify(0.0).current(0.0).previous(0.0),
                 Combination("02").when().previous(0.0).current(0.0).defuzzify(nan).current(def_cap).previous(0.0),
                 Combination("03").when().previous(0.0).current(0.0).defuzzify(inf).current(cap).previous(0.0),
-                Combination("04").when().previous(0.0).current(nan).defuzzify(0.0).current(0.0).previous(0.0
+                Combination("04").when().previous(0.0).current(nan).defuzzify(0.0).current(0.0).previous(
+                    0.0
                 ),  //.previous(nan),
-                Combination("05").when().previous(0.0).current(nan).defuzzify(nan).current(def_cap).previous(0.0
+                Combination("05").when().previous(0.0).current(nan).defuzzify(nan).current(def_cap).previous(
+                    0.0
                 ),  //.previous(nan),
-                Combination("06").when().previous(0.0).current(nan).defuzzify(inf).current(cap).previous(0.0
+                Combination("06").when().previous(0.0).current(nan).defuzzify(inf).current(cap).previous(
+                    0.0
                 ),  //.previous(nan),
                 Combination("07").when().previous(0.0).current(inf).defuzzify(0.0).current(0.0).previous(cap),
                 Combination("08").when().previous(0.0).current(inf).defuzzify(nan).current(def_cap).previous(cap),
@@ -1396,11 +1120,14 @@ namespace fuzzylite { namespace test { namespace output_variable {
                 Combination("19").when().previous(inf).current(0.0).defuzzify(0.0).current(0.0).previous(0.0),
                 Combination("20").when().previous(inf).current(0.0).defuzzify(nan).current(def_cap).previous(0.0),
                 Combination("21").when().previous(inf).current(0.0).defuzzify(inf).current(cap).previous(0.0),
-                Combination("22").when().previous(inf).current(nan).defuzzify(0.0).current(0.0).previous(inf
+                Combination("22").when().previous(inf).current(nan).defuzzify(0.0).current(0.0).previous(
+                    inf
                 ),  //.previous(nan),
-                Combination("23").when().previous(inf).current(nan).defuzzify(nan).current(def_cap).previous(inf
+                Combination("23").when().previous(inf).current(nan).defuzzify(nan).current(def_cap).previous(
+                    inf
                 ),  //.previous(nan),
-                Combination("24").when().previous(inf).current(nan).defuzzify(inf).current(cap).previous(inf
+                Combination("24").when().previous(inf).current(nan).defuzzify(inf).current(cap).previous(
+                    inf
                 ),  //.previous(nan),
                 Combination("25").when().previous(inf).current(inf).defuzzify(0.0).current(0.0).previous(cap),
                 Combination("26").when().previous(inf).current(inf).defuzzify(nan).current(def_cap).previous(cap),
@@ -1408,9 +1135,11 @@ namespace fuzzylite { namespace test { namespace output_variable {
             });
         }
 
-        SECTION("(e) When default is higher than the range, and locks previous value, "
-                "previous is used when defuzzification is nan"
-                "default is used when previous is nan") {
+        SECTION(
+            "(e) When default is higher than the range, and locks previous value, "
+            "previous is used when defuzzification is nan"
+            "default is used when previous is nan"
+        ) {
             const scalar def = 7.7;
             assert_that.variable->setDefaultValue(def);
             assert_that.variable->setLockValueInRange(false);
@@ -1421,44 +1150,58 @@ namespace fuzzylite { namespace test { namespace output_variable {
                 Combination("02").when().previous(0.0).current(0.0).defuzzify(nan).current(0.0).previous(0.0),
                 Combination("03").when().previous(0.0).current(0.0).defuzzify(inf).current(inf).previous(0.0),
                 // previous value won't update to non-finite current value
-                Combination("04").when().previous(0.0).current(nan).defuzzify(0.0).current(0.0).previous(0.0
+                Combination("04").when().previous(0.0).current(nan).defuzzify(0.0).current(0.0).previous(
+                    0.0
                 ),  //.previous(nan),
-                Combination("05").when().previous(0.0).current(nan).defuzzify(nan).current(0.0).previous(0.0
+                Combination("05").when().previous(0.0).current(nan).defuzzify(nan).current(0.0).previous(
+                    0.0
                 ),  //.previous(nan),
-                Combination("06").when().previous(0.0).current(nan).defuzzify(inf).current(inf).previous(0.0
+                Combination("06").when().previous(0.0).current(nan).defuzzify(inf).current(inf).previous(
+                    0.0
                 ),  //.previous(nan),
-                Combination("07").when().previous(0.0).current(inf).defuzzify(0.0).current(0.0).previous(0.0
+                Combination("07").when().previous(0.0).current(inf).defuzzify(0.0).current(0.0).previous(
+                    0.0
                 ),  //.previous(inf),
-                Combination("08").when().previous(0.0).current(inf).defuzzify(nan).current(0.0).previous(0.0
+                Combination("08").when().previous(0.0).current(inf).defuzzify(nan).current(0.0).previous(
+                    0.0
                 ),  //.previous(inf),
-                Combination("09").when().previous(0.0).current(inf).defuzzify(inf).current(inf).previous(0.0
+                Combination("09").when().previous(0.0).current(inf).defuzzify(inf).current(inf).previous(
+                    0.0
                 ),  //.previous(inf),
 
                 Combination("10").when().previous(nan).current(0.0).defuzzify(0.0).current(0.0).previous(0.0),
-                Combination("11").when().previous(nan).current(0.0).defuzzify(nan).current(0.0).previous(0.0
+                Combination("11").when().previous(nan).current(0.0).defuzzify(nan).current(0.0).previous(
+                    0.0
                 ),  //.current(def)
                 Combination("12").when().previous(nan).current(0.0).defuzzify(inf).current(inf).previous(0.0),
                 Combination("13").when().previous(nan).current(nan).defuzzify(0.0).current(0.0).previous(nan),
                 Combination("14").when().previous(nan).current(nan).defuzzify(nan).current(def).previous(nan),
                 Combination("15").when().previous(nan).current(nan).defuzzify(inf).current(inf).previous(nan),
                 // previous value won't update to non-finite current value
-                Combination("16").when().previous(nan).current(inf).defuzzify(0.0).current(0.0).previous(nan
+                Combination("16").when().previous(nan).current(inf).defuzzify(0.0).current(0.0).previous(
+                    nan
                 ),  //.previous(inf),
-                Combination("17").when().previous(nan).current(inf).defuzzify(nan).current(def).previous(nan
+                Combination("17").when().previous(nan).current(inf).defuzzify(nan).current(def).previous(
+                    nan
                 ),  //.previous(inf),
-                Combination("18").when().previous(nan).current(inf).defuzzify(inf).current(inf).previous(nan
+                Combination("18").when().previous(nan).current(inf).defuzzify(inf).current(inf).previous(
+                    nan
                 ),  //.previous(inf),
 
                 Combination("19").when().previous(inf).current(0.0).defuzzify(0.0).current(0.0).previous(0.0),
-                Combination("20").when().previous(inf).current(0.0).defuzzify(nan).current(0.0).previous(0.0
+                Combination("20").when().previous(inf).current(0.0).defuzzify(nan).current(0.0).previous(
+                    0.0
                 ),  //.current(inf)
                 Combination("21").when().previous(inf).current(0.0).defuzzify(inf).current(inf).previous(0.0),
                 // previous value won't update to non-finite current value
-                Combination("22").when().previous(inf).current(nan).defuzzify(0.0).current(0.0).previous(inf
+                Combination("22").when().previous(inf).current(nan).defuzzify(0.0).current(0.0).previous(
+                    inf
                 ),  //.previous(nan),
-                Combination("23").when().previous(inf).current(nan).defuzzify(nan).current(inf).previous(inf
+                Combination("23").when().previous(inf).current(nan).defuzzify(nan).current(inf).previous(
+                    inf
                 ),  //.previous(nan),
-                Combination("24").when().previous(inf).current(nan).defuzzify(inf).current(inf).previous(inf
+                Combination("24").when().previous(inf).current(nan).defuzzify(inf).current(inf).previous(
+                    inf
                 ),  //.previous(nan),
                 Combination("25").when().previous(inf).current(inf).defuzzify(0.0).current(0.0).previous(inf),
                 Combination("26").when().previous(inf).current(inf).defuzzify(nan).current(inf).previous(inf),
